+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 #### Discovery
+##### Version 1.1.0
+> 1. use kratos pkg
+> 2. replace gin by kratos/bm
+> 3. fix poll return nil when be canceled.
+> 4. add init protect mode
+> 5. supoort set.
+
 ##### Version 1.0.2
 > 1.fix nodesproc. get all zone nodes.